// src/app/store/todo-state/entity/todo.enum.ts
export  enum TodoEnum {
  GET_REQUEST = '[GET REQUEST] Todos',
  GET_RESPONSE = '[GET RESPONSE] Todos',
  PUT_REQUEST = '[PUT REQUEST] Todos',
  PUT_RESPONSE = '[PUT RESPONSE] Todos',
  COMPLETE_REQUEST = '[COMPLETE REQUEST] Todos',
  COMPLETE_RESPONSE = '[COMPLETE RESPONSE] Todos',
  POST_REQUEST = '[POST REQUEST] Todos',
  POST_RESPONSE = '[POST RESPONSE] Todos',
  DELETE_REQUEST = '[DELETE REQUEST] Todos',
  DELETE_RESPONSE = '[DELETE RESPONSE] Todos',
}



export enum ActionModel {
  POST = 'POST',
  PUT = 'PUT',
}